Fix various issues mentioned in code review
https://github.com/18F/identity/pull/2

* Add missing newline to session_store.rb
* Correct repo URLs in debug template
* Remove autogenerated js helper files
* Remove JS helpers for unused FAQ section
* Remove incorrect logo file
* Remove unused homepage.js file

* Add :code to list of filtered parameters
__Why__
The code param is used by the 2FA flow and does not need to be exposed in the logs.

__How__
Add it to list of filtered parameters.

* Remove unused CSP directive and customize config
__Why__
We do not need to support mixed content (HTTP/S) so I removed the relative directives and customized
the config to support only HTTPS content.

__How__
Remove the unused `preserve_schemas` directive and uncomment the `block_all_mixed_content`
directive.