Remove Omniauth logic about privileged user roles
**Why**:
- For now, we aren't planning on supporting Omniauth authorization
for privileged roles (admin and tech). This allows us to remove some
complexity from the code.

**How**:
- Remove references to groups
- Allow tech and admin users to sign in directly via the website
- Allow Omniauthed users to bypass 2FA and security questions setup.
This is needed because before, we were setting the role of the
Omniauthed user to either `tech` or `admin`, and those roles were
already able to bypass 2FA and security questions. But now, the role
is `user`, so we need to update the bypass logic.

**Additional improvements:**
- Refactor Omniauth-related code to adhere to Object-Oriented best
practices and make it easier to unit test
- Remove `kind` from `omniauth_callbacks` localized messages because
it's not easily generalizable to work with any third party provider