Move upaya codebase to repository root (#28)
__Why__
Not having the codebase in the repository root was causing problems with 3rd party services like
CodeClimate and Travis which expect the code at the root level. This also makes it easier for
deployment.

__How__
Move the code in code/upaya to / and update the .gitignore file. Commit the changes making sure not
to include any 'secret' files.