Refactor specs related to Active Job
**Why**:
- To make them more robust and less prone to intermittent failures

**How**:
- Instead of relying on a gem to check if a job was enqueued, check
that a specific method was called on an object, since those objects are
already unit tested separately
- Replace feature specs with controller and view specs
- Make sure Active Job queues are cleared **before** specs that
interact with them, as opposed to **after**.

For example, when a user signs in, they get sent an OTP via the
`send_two_factor_authentication_code` method. That method then
instantiates the `UserOtpSender` class, which contains logic that
determines which devices the OTP should be sent to. That logic is unit
tested separately in `otp_sender_spec`. So, instead of writing various
slow feature tests for all the different scenarios while signing in, we
just need to write one `SessionsController` test that verifies that
when a user signs in, the `send_two_factor_authentication_code` is
called on the current user.