Add support for TOTP soft tokens (e.g. Google Authenticator)
This change adds optional support for TOTP applications
such as Google Authenticator.  The options is currently
only available via the "edit profile" page and is not
presented to the user as part of the default sign-up flow.

If a user enables TOTP they will be asked for a TOTP token
during login, although they still have the option of
falling back to SMS.

Backup codes are not currently not supported.