Fix Capybara selector queries in view specs
**Why**: Some view specs were passing in the expected `h2` text as a
second parameter to `have_selector`, which is not what Capybara
expects. In version 2.7.0, Capyara introduced a warning when unused
parameters are passed to a selector query. In this case, the specs
weren't actually testing that the correct text was present, so the
tests were passing even when the text was not present.

**How**: Pass in a hash as the second parameter, with `text` as the key
and the expected text as the value.