Remove support for multiple second factors
This change removes support for OTP via email and multiple
second factors in general.  SMS is now the only method of OTP
delivery.

This simplifies the schema and removes the second_factors table.

Why:

OTP via email as second factor is bad idea since email is the
recovery mechanism for the first factor.

Supporting multiple seconds factors is nice, but second factors
are not generic enough to simply be driven by a database table.
Currently we only support SMS as a second factor so we can
keep the code simple by removing the addition database and
supporting code.