Raise error when keys are missing from secrets
**Why**:
When certain SAML-related keys are missing from secrets.yml, it causes
tests to fail, but without a clear explanation that the reason is due
to missing certs or RSA keys. A better developer experience would be
to raise an error right away during runtime if a required key is
missing, just like we've already been doing for Figaro keys.

**How**:
Override Figaro's `require_keys` method to also check secrets.yml, and
add the saml certs and RSA keys to the list of required keys.

An additional benefit of doing this is that we can get rid of the
`singular_account` fallback in `TwilioService` by adding
`twilio_accounts` to the list of required keys, and removing the
`twilio_account_sid`, `twilio_auth_token`, and `twilio_number` keys.