Don't use Twilio during performance testing mode
**Why**:
When `pt_mode` is on, SMS sending is disabled, and any OTP code is
considered valid. The way we were disabling SMS was by using Twilio's
Test Credentials, but those are meant for testing the API. Being able
to use the test Twilio client requires actual secret Twilio test
credentials in application.yml. This is overkill when we can prevent
an SMS from being sent in ways that don't invoke Twilio.

**How**:
Create a NullTwilioClient class that doesn't do anything, and call it
when pt_mode is on.

An added benefit is that we can update bin/setup so that the example
files are always copied over so that developers can make sure they
have the latest configuration. The only reason we used to prevent the
files from being overwritten was to allow developers to keep their
secret Twilio test credentials, but those are no longer needed.