Fix /sidekiq infinite redirect when signed out
Due to a bug in Devise, attempting to access the /sidekiq route while
signed out was causing an infinite redirect loop. This is because the
endpoint attempts to authenticate the user first to see if they are
an admin.

There is a pull request open that fixes this in Devise's FailureApp,
but since we are already overriding it, we can apply the fix directly.

Eventually, we might want to use basic HTTP auth instead of relying
on Devise and forcing developers to be both signed in and an admin.

Either way, this change is beneficial since it removes unecessary
conditionals in the FailureApp.