Remove password rotation feature
**Why**:
We have decided to not enforce password rotation at this time.
It is not required by the new NIST standards.

**How**:
Remove the devise_security_extension gem and related files.
The one thing we want to keep from the gem is regex validation
for the password, but since Rails provides built-in support for that,
we don't need the gem to implement that.