Remove pt_mode setting
**Why**: pt_mode (performance testing mode) is lagacy setting
that we don't currently use for performace testing but is
enabled in for development mode. It means that all OTPs are
considered valid which means its hard to verify the incorrect
OTP paths during development.

**How**: Introduce a more more meaningful "sms_disabled" setting
which, when used in developer mode, pre-fills the OTP values
rather than simply ignoring them.