Use named CSP override in SamlIdpController#auth
**Why**:
After a SAML-authed user finishes signing in, they get redirected to
the `acs_url` defined in `service_providers.yml` for the current
Service Provider in the current Rails environment. This redirect
happens via submitting a form in `saml_post_binding.html.slim`, which
is tied to the `auth` action of `SamlIdpController`.

Our default CSP config restricts form submissions to the current host,
but we need to whitelist approved URLs in this scenario.

**How**:
The `secure_headers` gem provides a way to define named overrides that
can then be called in controller actions. I've defined a `:saml`
override that parses `service_providers.yml` and adds all the
`acs_url`s for the current Rails env to the `form_action` directive.

Note that for the SamlTestController to redirect to `/api/saml/auth` in
specs, the `domain_name` in the test environment needs to be set to an
empty string. Hence, I've updated `application.yml.example`.