Move IdV-related attributes to Identities table
**Why**:
- We want to be able to track the service provider through which a
user achieved a particular LOA level.
- The `ial_token` field on the User model does not apply to this app
because we are using the SAML authn_context to determine whether or
not a user needs to verify their identity.

**How**:
- Remove legacy IdV-related fields and methods from the User model
- Update SamlIdpController to determine whether or not to redirect to
an IdV URL based on the LOA in the requested authn_context
- Add `id_proofers.yml` to map an LOA level to a particular ID proofer

Note that these are the smallest changes to support this feature.
Internal IdV is not implemented here.