Set type of code and mobile input fields to improve UX
__Why__
It did not automatically show the custom telephone input on the mobile field and
it required an extra step to get to the number input on the keyboard.

__How__
Set the code inputs to type `number` and mobile inputs to type `tel`. Update the
stylesheets to ignore the number field dial in desktop browsers for the code
input field.