Remove hard coded test cert fingerprints from app code
__Why__
It took me awhile to figure out why the tests were failing after updating the
certs. The fingerprint for the cert used by the test SP was hardcoded in the app
code in the test env.

__How__
Calculate the fingerprint the way it does for other service provider certs, as
it should with the test code.