Do not redirect users who sign in directly
**Why**: Previously, if a user signed in directly to the IdP app
while having an active Identity (any Identity where
`last_authenticated` is not `nil`), they would get redirected to the
Service Provider tied to that Identity. This was due to the logic in
the `SpRedirect` concern, which came from the save-ferris codebase.

We don't need this logic in the IdP app. When a user signs in directly,
they should remain on the IdP app.