Improve email confirmation UX (#182)
**Why**:
To address the following issues:

- When a user who is already confirmed clicks on the link in their
confirmation email again, they remain on the "Resend confirmation
instructions" screen with an error message that their email has already
been confirmed and they should try signing in. A better UX would be to
redirect them to the appropriate place depending on whether or not they
are signed in.

- If a user confirms, then signs out, and comes back to the site after
a while, forgetting they already confirmed, then clicks "Didn't receive
confirmation instructions?", and resends confirmation instructions to
that same email, they will never receive an email.

**How**:

- If a user is already confirmed and clicks the confirmation link again
while signed out, they are redirected to the sign in page, with the
following flash error message: `Your email address has already been
confirmed. Please sign in.`. If they are already signed in, they are
redirected to the dashboard (which in turn might redirect them to the
part of the account creation process they haven't completed yet).

- If a user is already confirmed and resends confirmation instructions
to their confirmed email, they will receive an email letting them know
their email is already confirmed, with a link to sign in or reset their
password if they forgot it.

- When a user first clicks the confirmation link, the flash notice says
`Your email address has been successfully confirmed. To continue
creating your account, please set a password.` as opposed to just
`Your email address has been successfully confirmed.`. This indicates
that the account is not fully created yet.

- The confirmation email now says `to continue creating your account`,
as opposed to `to finish creating ...` because clicking the confirmation
link does not finish creating the account.