Prevent unwanted emails sent to existing users
**Why**: After signing up with an email, each subsequent time you tried
to sign up with the same email, you would receive an email letting you
know the email is already in use. This means that anyone could try to
sign up with your email and keep spamming you.

Similary, I realized that a change I made recently would also allow
someone to spam you by continuously sending the confirmation
instructions to you after you had already confirmed your email.

**How**:
- Once a user registers, don't send any emails if they try to sign up
again with the same email. This is actually the default behavior in
Devise, but we were inadvertently altering it.

- Undo the sending of the `already_confirmed` email if someone tries
to send confirmation instructions to an existing user.