Store PII on its own table
**Why**:

PII is generated for every proofing session, and a user
may proof more than once. It is to the user's benefit
and convenience that PII is preserved, for
(a) passing to relying parties as required, and
(b) pre-populating forms when re-proofing.

**How**:

Per https://github.com/18F/identity-private/issues/479
PII is stored on its own table, with a one-to-many relationship
to users table.