Enable SAML encryption by default
**Why**: We want to enable encryption by default.  This
change also removes some un-used complexity around the
setting of key_transport and block_encryption.  We currently
only support single value for these setting so the simplest
option is just to hardcode them.

Disabling of saml encryption is done in application.yml
so that it can be overridden by the environment if
needed.