Store PII on its own table (#277)
* Add IdV tests

**Why**: IdV controllers lacked some test coverage,
particularly for failure cases.

**How**: Updates the identity-proofer gem to support
mock vendor failure triggers.

* Store PII on its own table

**Why**:

We collect PII for 2 reasons:

* identity resolution in proofing process,
* pass verified values to relying parties as required.

**How**:

Per https://github.com/18F/identity-private/issues/479
PII is stored on its own table, with a one-to-many relationship
to users table. The 'active' and 'verified' flags are tracked
in order to indicate the status of each profile record.