Add more events to track
**Why**: To have more complete analytics

Note that the refactoring of `PasswordsController` was necessary in
order to add event tracking without introducing new Rubocop and Reek
offenses.

Some tests were refactored as well to avoid slowing them down even more
by adding a new `it` block for the new tracking test. For tests that
use the database, it's best to combine all expectations in the same
`it` block to avoid hitting the DB more than once.