Use SimpleForm + ActiveModel::Model for IdV Form
**Why**: It keeps our forms consistent throughout the app and makes
them easier to change and maintain. By using `ActiveModel::Model`,
we can take advantage of the built-in Rails validations, and with
Simple Form, we can automatically display the errors inline next to
each field that has an error. Another issue this commit fixes:

- If the form was invalid, the user was redirected to the form with
flash messages. A redirect means that they will lose any values they
entered in the form, forcing them to enter everything all over again.
This is not the conventional approach. Instead, we use `render :index`
to render the page again, while preserving the valid values and
flagging the bad values, in a way that’s consistent with all our other
forms.