Move edit/email endpoint to its own controller
**Why**: To keep concerns separated. Each updatable attribute has its
own rules and logic. Having a separate controller for each attribute
makes it easier to make changes in the future, and prevents having
conditionals that will always return the same value.

For example, when updating a user's email, we know that their phone
number cannot change, so we don't need to ask whether the phone
number has changed each time. This conditional should only be asked
when updating the phone number.