Prevent TOTP-enabled users from viewing TOTP setup
**Why**: To avoid any potential confusion, we should disallow
TOTP-enabled users from accessing the TOTP setup page. Note that this
is an edge case since there is no direct link the TOTP setup page from
the app once they've enabled it, but they could potentially bookmark
the URL and visit it later. If they do, we redirect them to the profile
page.