Add throttle to limit new OTP requests
**Why**: Every time we send an OTP via SMS or voice, it costs us money.
To prevent someone maliciously trying to empty our pockets, I've
added a blocklist with Allow2Ban so that the user will be blocked from
making any more requests for 10 minutes after requesting 5 OTP codes
in the span of 5 minutes. Those three numbers are configurable via
`application.yml`.

Note that the throttle discriminator is the phone number as opposed to
the user since multiple users can have the same phone number.

The Allow2Ban filter requires an ActiveSupport-compliant cache store
that accepts an options hash as a second argument to the `read` method,
hence the need for the readthis gem.