Track GET request events via after_action (#389)
**Why**: To complement the events we track, which are mostly the
result of PUT requests. I was misusing Ahoy's `track_visit` method.
It's not meant to be used to track page views. Instead, we need to use
the same `track_event` method we've been using, and instead of placing
a `track_event` call in every controller, we can add an `after_action`
in `ApplicationController` and automatically track all GET requests.