Refactor IdV phone form (#426)
**Why**: For consistency and adherence to established patterns in the
app, controller actions should stick to RESTful actions, and forms that
pertain to a specific set of attributes and validations should interact
with a specific form object. Forms should use the simple_form gem for
consistency and reduced verbosity. This architecture makes it easier to
test, understand, and change the code.

**How**:
- Replace SessionsController#update_phone with a new Idv::PhoneController
that has `new` and `create` actions
- Use symbols instead of strings for `idv_params` keys
- Create a Form Object to handle the phone validation and idv_params
processing
- Use `simple_form_for` to be able to use the form object and simplify
the form.
- Make `idv_session` a public method since we need to be able to test
it. Private/protected methods are not meant to be tested directly, but
in this case, we need to be able to call it to make sure it gets set
properly.
- Remove SessionsController specs that were actually integration tests
and were already covered by feature specs. Add a feature spec for the
one controller test that wasn't already covered by a feature spec.
Controller tests are meant to be unit tests for a specific public
method. If a controller test calls multiple public methods in the same
test, it's not a unit test anymore, it's an integration test.
- Update feature specs to use valid phone numbers and to verify that
the phone displayed in the form is the formatted version of the phone
number.