Cleanup phone confirmation event tracking (#433)
**Why**: We were previously generating two different events
when the user confirmed their phone and they already had
one on record.  Instead we want to track a single event
which is either an initial confirmation or an update.

Also, cleanup event tracking for IdV phone confirmation
and remove the SMS message that we were sending during
the IdV process.  If we want to notify the user when
a proofing even occurs it should happen at the end of
the proofing session and should not be linked to the
phone number in particuar.