Refactor ServiceProvider and related code (#425)
* Refactor ServiceProvider and related code

**Why**: To simplify it and remove unused code. The main purpose of
this class is to provide a hash of service provider attributes based
on the contents of the `service_providers.yml` file. This means that
all we need to do is parse the file, symbolize the keys, process some
of the values, such as the cert, and generate a fingerprint from the
cert.

The only instance methods needed are `metadata`, which returns the
hash of attributes, and `encryption opts`. All the instance methods
based on the YAML keys that we used to have are not needed. They were
never called (or didn't need to be called) directly. I'm not sure what
I was thinking when I wrote that code. Note that `encryption_opts`
could be merged into the `metadata` hash as well, but it's much easier
to test when it's a separate method.

**How**:
- Remove all instance methods based on YAML file keys.
- To return the proper hash in the `metadata` method, parse the YAML
file, symbolize the keys, and merge any other attributes that aren't
read from the YAML file, such as the fingerprint.
- Make the `block_encryption` key required in the YAML file as opposed
to using a conditional in the code.
- Remove `metadata_url` entries from the YAML file since those are not
used. None of those URLs exist.
- Pass in the `OpenSSL::X509::Certificate` instance of the cert to the
Fingerprinter class, as opposed to creating two instances, one in the
ServiceProvider class, and one in the Fingerprinter class.
- Raise an error if the issuer is invalid (i.e not present in the YAML
file).