Add table to hold user visible account activity (#408)
**Why**: We want to give the user some visibility into
the history of the activity on their account.  The precise
subset of events and the retention schedule are not final
but this gives us a simple framework for managing them.

This is only the first part of this feature.  Still to come
is the ability to limit the number of events stored per user
and perhaps storing the location and user agent information
about the request that triggerd each event.