Speed up session timeout specs
**Why**: To shave a few seconds off our test suite

**How**:
- Use `stub_sign_in` to avoid creating a user in the DB in controller
specs that don't need it

- Check the request headers for the proper AJAX headers instead of
having a feature test with a sleep of 4 seconds. That spec was written
to test a bug that was due to not setting the proper AJAX headers.
If we can test that the headers are present, then we can be confident
the bug does not exist.