Remove need for domain entries in service_providers.yml
**Why**: The list of authorized service providers should be configurable
on each server, so that the live production server can have a different
list than our demo servers for example.

**How**: Per the Twelve Factor App principles, use ENV vars to allow
this kind of configuration. Hence, I've added a
`valid_service_providers` entry to `application.yml` that can be
customized on each server.

- Rename the `service_providers` initializer to `_service_providers`
to allow the `secure_headers` initializer to use the `SERVICE_PROVIDERS`
constant. Initializers get loaded alphabetically.

- Remove the `valid_hosts` key from `service_providers.yml` since it's
redundant.

- Add a `before_action` to SamlIdpController#auth to validate the
service provider, in case someone tries to send a SAML request with
a service provider that is not approved. Capture this event in our
analytics.