Add support for signing in with a recovery code
**Why**: To allow users who lose access to their 2FA phone to recover
access to their account.

**How**:
- Follow the design pattern of keeping each method of authentication in
its own controller, and a form object to validate the code.
- Generate a 16-character alphanumeric random code and one-way encrypt
it in the database. This has the advantage of keeping the code secure,
but the disadvantage that it cannot be displayed to the user more than
once, which could be problematic if they generate multiple codes and
don't remember which one is the latest one.
- When a user first signs up, after they confirm their 2FA phone, they
will be presented with their recovery code with instructions for how
and where to keep it safe.
- Once signed in, a user can generate a new code. Note that this
currently doesn't include any messages about what that implies.
- When a user signs in with their recovery code, a new one is
immediately generated, which replaces their previous one.