Handle multiple ways to 2FA in own controllers (#473)
**Why**: So that the two_factor_authentication_controller is not
responsible for all the different cases and to adhere to the Single
Responsibility Principle.

**How**:
- Create 2 new controllers for processing codes entered by the user:
one that verifies OTP codes, and one that verifies TOTP codes.
- The `TwoFactorAuthentication` controller no longer handles codes
entered by the user. It is now only responsible for sending an OTP
to the user based on the chosen delivery method.
- Extract the common code shared by the OTP and TOTP verification
controllers to a concern called `TwoFactorAuthenticatable`
- Create new routes with more descriptive path names, such as
`login/two-factor/sms`
- Add a form object to ensure that only allowed OTP delivery methods
are handled. If an invalid method is submitted, the user gets
redirected to the screen where they must choose between SMS and Voice
delivery.
- Instead of setting various instance variables to keep track of the
selected delivery method, use the current URL path since the new
endpoints correspond to the delivery method.