Update session timeout text
**Why**: To reduce confusion in the scenario where the session times out
before a user has authed via 2FA. Our app is designed such that a user
is considered signed in (by the code) after they enter a valid email
and password, but a user may not think they are logged in until they
go through 2FA.

Previously, the timeout message referred to the user being logged out,
which technically is correct, but may not make sense to the user. To
address this, I've updated the wording to refer to the session ending.