Small formatting stuff
**Why**: I just flipped through most of the view templates and saw that
we had extra newlines in some places and my natural instinct was to
remove them.

I also have my dotfiles set up to show a line at 80 chars, but I
realized after a bit that the limit on this proj is 100 so I made that
clearer in our rubocop config file.

The third small change is to put called methods below the methods in
which they are called. There is a name for that, but I am not able to
recall it at this moment...