Reduce duplication in OTP verification contexts (#584)
* Reduce duplication in OTP verification contexts

**Why**: To make the code less complex, easier to read, easier to
change, and to reduce the chances of breaking one part of the code when
modifying another.

**How**:
- Think of OTP delivery as a separate service that always acts
the same way regardless of context. Whether you are authenticating,
or confirming a new phone, or updating an existing phone, or confirming
a phone different from your existing phone when proofing, you are
requesting an OTP to be sent to you. This sending of the OTP should
act the same in all scenarios, and there should only be one controller
that performs this action, not several separate ones for each scenario.

Therefore, always use Devise::TwoFactorAuthenticationController#show for
the delivery selection, and always redirect to OtpVerificationController
to process the code entered by the user.

- This means that instead of generating and storing the OTP code in the
session in the phone confirmation context, we use
`User#create_direct_otp` in all contexts, which means that the concept
of resending the same code no longer exists. If you didn't receive the
code, the link will always say "Send a new one", and a new code will be
generated and sent. This makes things consistent and means we can be
confident that `direct_otp_sent_at` will always be present in the
OtpSender jobs.

- To differentiate between what happens after the user enters a valid
code, we introduce a context parameter, and add logic to handle each
context. This can certainly be improved, but is a good first step for
now IMO.

- Use Form Objects for OTP and TOTP verification to follow the pattern
we've been using to simplify analytics calls. There's more refactoring
to be done to clean up analytics, but that will be in another PR.

- I introduced some Rubocop offenses that will be fixed in a subsequent
PR.

* Address PR feedback

* Fix confirmation context bugs

- In confirmation context, OTP attempts should not be limited
- In IdV context, reenter phone number link should point to phone screen

* More spec tweaks

* Use let instead of instance variable