Only handle phone change during confirmation context
**Why**: If a user has an existing phone number, then sets up and
confirms a different phone number during IdV, handling this as a phone
change event will send an SMS to the existing phone number alerting the
user that their phone number was changed, which isn't true. It will also
log analytics related to changing the user's phone, which is also
inaccurate.

**How**: Add another condition to the if statement that checks that the
context is confirmation in order to handle a phone change event.