Allow multiple invalid password submissions (#725)
**Why**: To fix a bug I introduced when I refactored analytics.

**How**: Instead of using a completely different Devise method,
use the same one Devise uses by default in this context, but modify
it so that it doesn't update the user's password. This is because
we removed the model validations, so we shouldn't attempt to update
the user's password until the form has been processed, i.e. until
we know whether or not the password is valid.

Devise, on the other hand, saves the password as soon as `#update` is
called, then checks to see if the resource has any errors. Since we only
validate the password via the form object, we can't update the user
until we know the password is valid.

Note that the original Devise method is a class method on User called
`reset_password_by_token`, but I made it a controller method and
renamed it `user_matching_token` since it better describes what it
returns.