Adds information about overcommit (#724)
**Why**:
The overcommit tool requires that a signature of the current file be
generated before git commands can be used when developing this
application. Although simple to set up, it isn't clear what overcommit
is or why the signature is necessary if one hasn't used it before.

A small section on overcommit is added to the README so new devs
can be made aware of it.