Improve app and events logging
**Why**: The logstash-logger gem was automatically formatting Rails server
and console messages in JSON format, and doesn't provide an option to
keep the original logging as it is. We're not currently using logstash,
so we can revisit how to send logs to logstash later if needed.

**How**:
- Remove the logstash-logger gem, and create a new custom logger for
analytics that logs events to a separate file.
- Remove the Procfile pipe since it's no longer needed. Lograge knows
not to format SQL logs, and Rails console logs.
- Add an example logstash.conf to allow running analytics queries
locally