Redirects users from /users to /users/sign_up
**Why**:
When a user sits on the users/sign_up page for too long, they are
directed to the /users page, which is not defined. Instead, redirect
them back to the sign_up page.

Adds test for redirect from /users

**Why**:
Tests are important for determining if the app is behaving properly

Updates capybara specs

**Why**:
We need to test that the modal redirects the user to the same page

Removes redirect from /users

**Why**:
Providing a route to /users doesn't satisfy the issue this PR is meant
to resolve

Uses capybara matcher for current path

**Why**:
Using have_current_path means the code doesnt have to have workarounds
like sleep statements