Refactor 2FA Max attempts analytics
**Why**: We were incorrectly reporting a Max Attempts Reached event
when the user attempted to sign in while locked out. The event should
only be triggered when the user enters an incorrect 2FA code too many
times.

**How**:
- Remove `verify_user_is_not_second_factor_locked` from
`TwoFactorAuthenticatable`, and put that logic in SessionsController,
since a user should not get past the sign in screen if locked out.
- Add a property to the signin analytics to show if the user is
currently locked out or not.