Allow user to stay on account locked page
**Why**: For a better UX, we want to leave the user on the account
locked page for the duration of the lockout period.

**How**:
- Reverse the order of events in `handle_second_factor_locked_user`.
The reason is that application.html.slim wasn't called after signing
the user out, and therefore the timeout JS was still active.
By reversing the order, we force a new call to the layout template,
which then sees that a current user does not exist, and therefore does
not trigger the timeout JS.
- Revert the change from `current_user` to `user_fully_authenticated?`
in the application layout to allow users who only sign in with email and
password to see the time out warning modal.