Sign user out when reached max password attempts
**Why**:
If a user is already signed in, and is trying to change their password,
email, or phone, they will be prompted to enter their current password
before they can make the changes. If an incorrect password is being
entered repeatedly, it is most likely due to a malicious person taking
over the user's machine that they left unattended. In that case, for the
user's security, we sign the user out.

The max attempts count is configurable, and has been set to 3 for now,
to match the OTP max attempts.