Improve Password Reset UX
**Why**: When a user goes to reset their password, after they submit
their email, they were sent back to the sign in page with a flash
message. This is not consistent with what happens when you submit your
email when creating an account, i.e. you are given a chance to send
the email again, along with detailed instructions.

This PR makes the experience of resetting a password consistent with
the account creation.

**How**:
- In order to prevent revealing whether or not an email exists in our
database, we must present the same password reset experience regardless
of the scenario and state of the user. This means that we should always
send the password reset email, even if the user is unconfirmed. This
means we only need to show the new
`app/views/forgot_password/show.html.slim` page that I added in this PR.

- Related to the above, since we're now sending password reset emails to
unconfirmed users, we confirm them after resetting their password, since
we've essentially confirmed that they own that email address. It doesn't
matter whether the confirmation comes via a confirmation email or a
password reset email.

- Remove the `ValidEmailParameter` concern from
`ResetPasswordsController`, and use a form object instead with our
existing `FormEmailValidator`, to keep email validation consistent
everywhere, and for better analytics. The concern was not including the
domains in our banned list (see `valid_email.yml`).

- Move the title and headings localizations of the `sign_up/emails/show`
view to a top-level key since this will be the same across all views
that are a result of submitting an email field that causes an email to
be sent, including the new forgot password view that was added here.

- Replace feature specs for edge cases with controller specs