Raise error if can't decrypt attribute w/ any key
**Why**: If for some reason, the `attribute_encryption_key` was rotated,
but the `attribute_encryption_key_queue` was not updated to include the
previous key, then we want to make sure to raise an error to prevent
the following scenario that happened during the most recent deployment
to dev: Any existing user who signed in after the deployment had their
email reset to an empty string, preventing them from ever being able to
sign back in.

This was due to `EncryptedAttribute#try_decrypt_with_all_keys` returning
the array of keys in `attribute_encryption_key_queue` when it was not
able to decrypt with any of the keys.

**How**:
- If no valid decryption keys are found, raise an error. Do this by
changing `EncryptedAttribute#try_decrypt_with_all_keys` so that it
uses `each_with_object` to return an empty string if no keys match.
- Add a unit test for the scenario where the queue does not contain any
previous keys.
- Add an integration test to make sure the User's email does not get
set to an empty string.
- Update the example config documentation to explain that the queue must
always contain at least the current key, and should contain the current
key by default.